Implemented selectFilter in angular 2
The component for select filter mostly just passes its configuration down to the select component. The only behavior that it adds is setting the value on the filter and refreshing the data source.
