+++ source/components/cardContainer/filters/selectFilter/selectFilter.ng1.html
@@ -1,0 +1,4 @@
+<div class="filter-group">
+	<rl-select ng-model="filter.selectedValue" options="filter.options" label="{{filter.label}}"
+			   transform="filter.transform" get-options="filter.getOptions()" null-option="{{filter.nullOption}}" item-as="{{filter.itemAs}}" template="filter.template"></rl-select>
+</div>
